[#62] refactor(files): moved all blog components to blog catalog
